+++ CHANGES.md
@@ -3,7 +3,9 @@
 Unreleased
 
 -   The description for a model's object, and for an attribute's field and
-    argument, is set from their docstrings. :issue:`19`
+    argument, is set from their docstrings. {issue}`19`
+-   Handle fragments when inspecting query to load relationships. {issue}`21`
+-   Clearer error when SQLAlchemy session is not passed in GraphQL context.
 
 
 ## Version 1.0.0